LPS-102252 Moves to display context
